Move contributor components
